requirements.txt
flask
numpy 
pandas 
seaborn 
scikit-learn
ipykernel
pymongo
imbalanced-learn
dill
python-dotenv
xgboost
pyyaml